Floodstation service

+++ api/src/main/java/com/safetynet/api/controller/dto/StationDto.java
@@ -2,4 +2,4 @@ package com.safetynet.api.controller.dto;
 
 import java.util.List;
 
-public record StationDto(List<PersonDto> person, int child, int adult){}
+public record StationDto(List<PersonDto> person, Long child, Long adult){}
